--- a/src/main/java/org/jboss/threads/EnhancedQueueExecutorBase4.java
+++ b/src/main/java/org/jboss/threads/EnhancedQueueExecutorBase4.java
@@ -26,7 +26,7 @@ abstract class EnhancedQueueExecutorBase4 extends EnhancedQueueExecutorBase3 {
      * Padding fields.
      */
     @SuppressWarnings("unused")
-    int p00, p01, p02, p03,
+    long p00, p01, p02, p03,
         p04, p05, p06, p07,
         p08, p09, p0A, p0B,
         p0C, p0D, p0E, p0F;