Intel CPUs uses an optimization called "Adjacent line prefetchers" that fetch 2 cache lines while accessing the first one, hence is safer to use 128 bytes padding between hot accessed fields to avoid false sharing or just noisy neighbours.
I don't have yet access to a big box to run some tests, but that's why `-XX:ContendedPaddingWidth` is 128 by default.