@carterkozak 2020-04-13
I put together a quick `perf c2c` jmh profiler, if you have a particular benchmark configuration you'd like to to try I'd be happy to a/b test.