@franz1981 2020-04-13
@carterkozak many thanks! Yep, probably will try 2 cases:

- one that would stress the `tail` vs `head` false sharing ie it's fine with just 1 producer and 1 consumer, possibly with different burst size (1, 10, 100): using `delay.c` == 0 will make consuming to often see an empty queue (that's good)
- another  with many producers/consumers (the default I assume)

My concern is that many of these test cases makes the producers/consumers to pay the cost to awake/fall asleep at some point and it could make vert complex to measure the gains, so any idea/proposal to bench the contention is wellcome :)