@franz1981 2020-04-13
one sec, I see that `PoolThreadNode` need the same..