@franz1981 2020-04-13
The post-padding is good to save from "noisy neighbours" on the heap, but is not necessary.