@carterkozak 2020-04-13
I'm not sure how helpful the batches are for overall benchmarking, they create the contention we're looking for, but I think tasks are completing faster than we're able to submit them. I added some debug logging locally after 100 tasks are submitted and found that between zero and one tasks were executing and the queue was empty, even with 32 submitting threads.

Pushed the profiler implementation here: https://github.com/carterkozak/perf-c2c-jmh-profiler

# one that would stress the tail vs head false sharing ie it's fine with just 1 producer and 1 consumer (and with 2 cores), possibly with different burst size (1, 10, 100): using delay.c == 0 will make consuming to often see an empty queue (that's good)

Before:
```
# JMH version: 1.22
# VM version: JDK 14-ea, OpenJDK 64-Bit Server VM, 14-ea+36-Ubuntu-2
# VM invoker: /usr/lib/jvm/java-14-openjdk-amd64/bin/java
# VM options: -Xmx2g -Xms2g -XX:+UseParallelOldGC -XX:-UseBiasedLocking
# Warmup: 2 iterations, 7 s each
# Measurement: 4 iterations, 7 s each
# Timeout: 10 min per iteration
# Threads: 1 thread, will synchronize iterations
# Benchmark mode: Throughput, ops/time
# Benchmark: net.ckozak.ExecutorBenchmarks.benchmarkSubmit
# Parameters: (burstSize = 1, cores = 2, executorThreads = 1,1, factory = EQE_SPIN_LOCK)

# Run progress: 0.00% complete, ETA 00:02:06
# Fork: 1 of 1
# Preparing profilers: LinuxPerfCacheToCacheProfiler 
# Profilers consume stdout and stderr from target VM, use -v EXTRA to copy to console
OpenJDK 64-Bit Server VM warning: Option UseParallelOldGC was deprecated in version 14.0 and will likely be removed in a future release.
OpenJDK 64-Bit Server VM warning: Option UseParallelOldGC was deprecated in version 14.0 and will likely be removed in a future release.
WARNING: An illegal reflective access operation has occurred
WARNING: Illegal reflective access by org.openjdk.jmh.util.Utils (file:/home/ckozak/.gradle/caches/modules-2/files-2.1/org.openjdk.jmh/jmh-core/1.22/ce10949eab30c7378f73ec0cf116ad2c3cc2be80/jmh-core-1.22.jar) to field java.io.PrintStream.charOut
WARNING: Please consider reporting this to the maintainers of org.openjdk.jmh.util.Utils
WARNING: Use --illegal-access=warn to enable warnings of further illegal reflective access operations
WARNING: All illegal access operations will be denied in a future release
Apr 13, 2020 1:33:53 PM org.jboss.threads.Version <clinit>
INFO: JBoss Threads version 3.1.0.Final
# Warmup Iteration   1: 187587.394 ops/s
# Warmup Iteration   2: 191139.283 ops/s
Iteration   1: 148161.721 ops/s
Iteration   2: 142321.919 ops/s
Iteration   3: 138902.430 ops/s
Iteration   4: 143115.144 ops/s
# Processing profiler results: LinuxPerfCacheToCacheProfiler 


Result "net.ckozak.ExecutorBenchmarks.benchmarkSubmit":
  143125.303 ±(99.9%) 24703.539 ops/s [Average]
  (min, avg, max) = (138902.430, 143125.303, 148161.721), stdev = 3822.900
  CI (99.9%): [118421.764, 167828.843] (assumes normal distribution)

Secondary result "net.ckozak.ExecutorBenchmarks.benchmarkSubmit:·perf-c2c":
Perf C2C report (stats):
--------------------------------------------------
=================================================
            Trace Event Information              
=================================================
  Total records                     :     402444
  Locked Load/Store Operations      :      25947
  Load Operations                   :     188379
  Loads - uncacheable               :          0
  Loads - IO                        :          0
  Loads - Miss                      :          2
  Loads - no mapping                :          0
  Load Fill Buffer Hit              :     111672
  Load L1D hit                      :      46057
  Load L2D hit                      :       3114
  Load LLC hit                      :      27485
  Load Local HITM                   :      26305
  Load Remote HITM                  :          0
  Load Remote HIT                   :          0
  Load Local DRAM                   :         49
  Load Remote DRAM                  :          0
  Load MESI State Exclusive         :          0
  Load MESI State Shared            :         49
  Load LLC Misses                   :         49
  LLC Misses to Local DRAM          :      100.0%
  LLC Misses to Remote DRAM         :        0.0%
  LLC Misses to Remote cache (HIT)  :        0.0%
  LLC Misses to Remote cache (HITM) :        0.0%
  Store Operations                  :     214065
  Store - uncacheable               :          0
  Store - no mapping                :          0
  Store L1D Hit                     :     206785
  Store L1D Miss                    :       7280
  No Page Map Rejects               :     306266
  Unable to parse data source       :          0

=================================================
    Global Shared Cache Line Event Information   
=================================================
  Total Shared Cache Lines          :       2957
  Load HITs on shared lines         :     151068
  Fill Buffer Hits on shared lines  :     107293
  L1D hits on shared lines          :      16561
  L2D hits on shared lines          :        158
  LLC hits on shared lines          :      27054
  Locked Access on shared lines     :      12281
  Store HITs on shared lines        :      90597
  Store L1D hits on shared lines    :      83551
  Total Merged records              :     116902

=================================================
                 c2c details                     
=================================================
  Events                            : cpu/mem-loads,ldlat=30/P
                                    : cpu/mem-stores/P
                                    : dummy:HG
  Cachelines sort on                : Total HITMs
  Cacheline data grouping           : offset,iaddr


# JMH version: 1.22
# VM version: JDK 14-ea, OpenJDK 64-Bit Server VM, 14-ea+36-Ubuntu-2
# VM invoker: /usr/lib/jvm/java-14-openjdk-amd64/bin/java
# VM options: -Xmx2g -Xms2g -XX:+UseParallelOldGC -XX:-UseBiasedLocking
# Warmup: 2 iterations, 7 s each
# Measurement: 4 iterations, 7 s each
# Timeout: 10 min per iteration
# Threads: 1 thread, will synchronize iterations
# Benchmark mode: Throughput, ops/time
# Benchmark: net.ckozak.ExecutorBenchmarks.benchmarkSubmit
# Parameters: (burstSize = 10, cores = 2, executorThreads = 1,1, factory = EQE_SPIN_LOCK)

# Run progress: 33.33% complete, ETA 00:01:28
# Fork: 1 of 1
# Preparing profilers: LinuxPerfCacheToCacheProfiler 
# Profilers consume stdout and stderr from target VM, use -v EXTRA to copy to console
Apr 13, 2020 1:34:37 PM org.jboss.threads.Version <clinit>
INFO: JBoss Threads version 3.1.0.Final
# Warmup Iteration   1: 134266.957 ops/s
# Warmup Iteration   2: 138051.287 ops/s
Iteration   1: 115648.736 ops/s
Iteration   2: 108865.262 ops/s
Iteration   3: 109362.413 ops/s
Iteration   4: 109866.042 ops/s
# Processing profiler results: LinuxPerfCacheToCacheProfiler 


Result "net.ckozak.ExecutorBenchmarks.benchmarkSubmit":
  110935.613 ±(99.9%) 20475.033 ops/s [Average]
  (min, avg, max) = (108865.262, 110935.613, 115648.736), stdev = 3168.534
  CI (99.9%): [90460.581, 131410.646] (assumes normal distribution)

Secondary result "net.ckozak.ExecutorBenchmarks.benchmarkSubmit:·perf-c2c":
Perf C2C report (stats):
--------------------------------------------------
=================================================
            Trace Event Information              
=================================================
  Total records                     :     407227
  Locked Load/Store Operations      :      47108
  Load Operations                   :     192247
  Loads - uncacheable               :          0
  Loads - IO                        :          0
  Loads - Miss                      :          1
  Loads - no mapping                :          4
  Load Fill Buffer Hit              :      95073
  Load L1D hit                      :      68231
  Load L2D hit                      :       1224
  Load LLC hit                      :      27668
  Load Local HITM                   :      26277
  Load Remote HITM                  :          0
  Load Remote HIT                   :          0
  Load Local DRAM                   :         46
  Load Remote DRAM                  :          0
  Load MESI State Exclusive         :          0
  Load MESI State Shared            :         46
  Load LLC Misses                   :         46
  LLC Misses to Local DRAM          :      100.0%
  LLC Misses to Remote DRAM         :        0.0%
  LLC Misses to Remote cache (HIT)  :        0.0%
  LLC Misses to Remote cache (HITM) :        0.0%
  Store Operations                  :     214980
  Store - uncacheable               :          0
  Store - no mapping                :          0
  Store L1D Hit                     :     209968
  Store L1D Miss                    :       5012
  No Page Map Rejects               :     227394
  Unable to parse data source       :          0

=================================================
    Global Shared Cache Line Event Information   
=================================================
  Total Shared Cache Lines          :       5478
  Load HITs on shared lines         :     137211
  Fill Buffer Hits on shared lines  :      85775
  L1D hits on shared lines          :      24254
  L2D hits on shared lines          :        100
  LLC hits on shared lines          :      27080
  Locked Access on shared lines     :      19588
  Store HITs on shared lines        :      81912
  Store L1D hits on shared lines    :      77195
  Total Merged records              :     108189

=================================================
                 c2c details                     
=================================================
  Events                            : cpu/mem-loads,ldlat=30/P
                                    : cpu/mem-stores/P
                                    : dummy:HG
  Cachelines sort on                : Total HITMs
  Cacheline data grouping           : offset,iaddr


# JMH version: 1.22
# VM version: JDK 14-ea, OpenJDK 64-Bit Server VM, 14-ea+36-Ubuntu-2
# VM invoker: /usr/lib/jvm/java-14-openjdk-amd64/bin/java
# VM options: -Xmx2g -Xms2g -XX:+UseParallelOldGC -XX:-UseBiasedLocking
# Warmup: 2 iterations, 7 s each
# Measurement: 4 iterations, 7 s each
# Timeout: 10 min per iteration
# Threads: 1 thread, will synchronize iterations
# Benchmark mode: Throughput, ops/time
# Benchmark: net.ckozak.ExecutorBenchmarks.benchmarkSubmit
# Parameters: (burstSize = 100, cores = 2, executorThreads = 1,1, factory = EQE_SPIN_LOCK)

# Run progress: 66.67% complete, ETA 00:00:44
# Fork: 1 of 1
# Preparing profilers: LinuxPerfCacheToCacheProfiler 
# Profilers consume stdout and stderr from target VM, use -v EXTRA to copy to console
Apr 13, 2020 1:35:21 PM org.jboss.threads.Version <clinit>
INFO: JBoss Threads version 3.1.0.Final
# Warmup Iteration   1: 26064.736 ops/s
# Warmup Iteration   2: 27610.142 ops/s
Iteration   1: 27528.567 ops/s
Iteration   2: 27304.186 ops/s
Iteration   3: 28262.288 ops/s
Iteration   4: 28296.499 ops/s
# Processing profiler results: LinuxPerfCacheToCacheProfiler 


Result "net.ckozak.ExecutorBenchmarks.benchmarkSubmit":
  27847.885 ±(99.9%) 3274.974 ops/s [Average]
  (min, avg, max) = (27304.186, 27847.885, 28296.499), stdev = 506.806
  CI (99.9%): [24572.911, 31122.859] (assumes normal distribution)

Secondary result "net.ckozak.ExecutorBenchmarks.benchmarkSubmit:·perf-c2c":
Perf C2C report (stats):
--------------------------------------------------
=================================================
            Trace Event Information              
=================================================
  Total records                     :     412372
  Locked Load/Store Operations      :      94154
  Load Operations                   :     195826
  Loads - uncacheable               :          0
  Loads - IO                        :          0
  Loads - Miss                      :          1
  Loads - no mapping                :          2
  Load Fill Buffer Hit              :      48892
  Load L1D hit                      :     118826
  Load L2D hit                      :        618
  Load LLC hit                      :      27421
  Load Local HITM                   :      24521
  Load Remote HITM                  :          0
  Load Remote HIT                   :          0
  Load Local DRAM                   :         66
  Load Remote DRAM                  :          0
  Load MESI State Exclusive         :          0
  Load MESI State Shared            :         66
  Load LLC Misses                   :         66
  LLC Misses to Local DRAM          :      100.0%
  LLC Misses to Remote DRAM         :        0.0%
  LLC Misses to Remote cache (HIT)  :        0.0%
  LLC Misses to Remote cache (HITM) :        0.0%
  Store Operations                  :     216546
  Store - uncacheable               :          0
  Store - no mapping                :          0
  Store L1D Hit                     :     214330
  Store L1D Miss                    :       2216
  No Page Map Rejects               :      60813
  Unable to parse data source       :          0

=================================================
    Global Shared Cache Line Event Information   
=================================================
  Total Shared Cache Lines          :      11722
  Load HITs on shared lines         :      89331
  Fill Buffer Hits on shared lines  :      27186
  L1D hits on shared lines          :      36168
  L2D hits on shared lines          :         39
  LLC hits on shared lines          :      25935
  Locked Access on shared lines     :      33649
  Store HITs on shared lines        :      52306
  Store L1D hits on shared lines    :      50845
  Total Merged records              :      76827

=================================================
                 c2c details                     
=================================================
  Events                            : cpu/mem-loads,ldlat=30/P
                                    : cpu/mem-stores/P
                                    : dummy:HG
  Cachelines sort on                : Total HITMs
  Cacheline data grouping           : offset,iaddr


# Run complete. Total time: 00:02:12

REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
experiments, perform baseline and negative tests that provide experimental control, make sure
the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
Do not assume the numbers tell you what you want them to tell.

Benchmark                                     (burstSize)  (cores)  (executorThreads)      (factory)   Mode  Cnt       Score       Error  Units
ExecutorBenchmarks.benchmarkSubmit                      1        2                1,1  EQE_SPIN_LOCK  thrpt    4  143125.303 ± 24703.539  ops/s
ExecutorBenchmarks.benchmarkSubmit:·perf-c2c            1        2                1,1  EQE_SPIN_LOCK  thrpt              NaN               #/op
ExecutorBenchmarks.benchmarkSubmit                     10        2                1,1  EQE_SPIN_LOCK  thrpt    4  110935.613 ± 20475.033  ops/s
ExecutorBenchmarks.benchmarkSubmit:·perf-c2c           10        2                1,1  EQE_SPIN_LOCK  thrpt              NaN               #/op
ExecutorBenchmarks.benchmarkSubmit                    100        2                1,1  EQE_SPIN_LOCK  thrpt    4   27847.885 ±  3274.974  ops/s
ExecutorBenchmarks.benchmarkSubmit:·perf-c2c          100        2                1,1  EQE_SPIN_LOCK  thrpt              NaN               #/op
```

After:
```
# JMH version: 1.22
# VM version: JDK 14-ea, OpenJDK 64-Bit Server VM, 14-ea+36-Ubuntu-2
# VM invoker: /usr/lib/jvm/java-14-openjdk-amd64/bin/java
# VM options: -Xmx2g -Xms2g -XX:+UseParallelOldGC -XX:-UseBiasedLocking
# Warmup: 2 iterations, 7 s each
# Measurement: 4 iterations, 7 s each
# Timeout: 10 min per iteration
# Threads: 1 thread, will synchronize iterations
# Benchmark mode: Throughput, ops/time
# Benchmark: net.ckozak.ExecutorBenchmarks.benchmarkSubmit
# Parameters: (burstSize = 1, cores = 2, executorThreads = 1,1, factory = EQE_SPIN_LOCK)

# Run progress: 0.00% complete, ETA 00:02:06
# Fork: 1 of 1
# Preparing profilers: LinuxPerfCacheToCacheProfiler 
# Profilers consume stdout and stderr from target VM, use -v EXTRA to copy to console
OpenJDK 64-Bit Server VM warning: Option UseParallelOldGC was deprecated in version 14.0 and will likely be removed in a future release.
OpenJDK 64-Bit Server VM warning: Option UseParallelOldGC was deprecated in version 14.0 and will likely be removed in a future release.
WARNING: An illegal reflective access operation has occurred
WARNING: Illegal reflective access by org.openjdk.jmh.util.Utils (file:/home/ckozak/.gradle/caches/modules-2/files-2.1/org.openjdk.jmh/jmh-core/1.22/ce10949eab30c7378f73ec0cf116ad2c3cc2be80/jmh-core-1.22.jar) to field java.io.PrintStream.charOut
WARNING: Please consider reporting this to the maintainers of org.openjdk.jmh.util.Utils
WARNING: Use --illegal-access=warn to enable warnings of further illegal reflective access operations
WARNING: All illegal access operations will be denied in a future release
Apr 13, 2020 1:29:55 PM org.jboss.threads.Version <clinit>
INFO: JBoss Threads version 3.1.0.Final
# Warmup Iteration   1: 190966.254 ops/s
# Warmup Iteration   2: 197949.166 ops/s
Iteration   1: 151416.521 ops/s
Iteration   2: 148418.446 ops/s
Iteration   3: 147975.967 ops/s
Iteration   4: 145744.934 ops/s
# Processing profiler results: LinuxPerfCacheToCacheProfiler 


Result "net.ckozak.ExecutorBenchmarks.benchmarkSubmit":
  148388.967 ±(99.9%) 15075.713 ops/s [Average]
  (min, avg, max) = (145744.934, 148388.967, 151416.521), stdev = 2332.983
  CI (99.9%): [133313.254, 163464.680] (assumes normal distribution)

Secondary result "net.ckozak.ExecutorBenchmarks.benchmarkSubmit:·perf-c2c":
Perf C2C report (stats):
--------------------------------------------------
=================================================
            Trace Event Information              
=================================================
  Total records                     :     402641
  Locked Load/Store Operations      :      27603
  Load Operations                   :     188232
  Loads - uncacheable               :          0
  Loads - IO                        :          0
  Loads - Miss                      :          1
  Loads - no mapping                :          2
  Load Fill Buffer Hit              :     111631
  Load L1D hit                      :      47497
  Load L2D hit                      :       2414
  Load LLC hit                      :      26649
  Load Local HITM                   :      25820
  Load Remote HITM                  :          0
  Load Remote HIT                   :          0
  Load Local DRAM                   :         38
  Load Remote DRAM                  :          0
  Load MESI State Exclusive         :          0
  Load MESI State Shared            :         38
  Load LLC Misses                   :         38
  LLC Misses to Local DRAM          :      100.0%
  LLC Misses to Remote DRAM         :        0.0%
  LLC Misses to Remote cache (HIT)  :        0.0%
  LLC Misses to Remote cache (HITM) :        0.0%
  Store Operations                  :     214409
  Store - uncacheable               :          0
  Store - no mapping                :          0
  Store L1D Hit                     :     205942
…